New implementation of sorting (using arrays)

+++ OpenMp_Task6/numeric.cpp
@@ -1,13 +1,13 @@
 #include "numeric.h"
 
 /**
- * Allocate array of double numver and initialize them by
- * random value
- * @param min The minimum value of number
- * @param max The maximum value of number
- * @param num The numbers of double number
- * @return The pointer to array of double number
- */
+* Allocate array of double numver and initialize them by
+* random value
+* @param min The minimum value of number
+* @param max The maximum value of number
+* @param num The numbers of double number
+* @return The pointer to array of double number
+*/
 double* OmpNumRandomGenerate(double min, double max, uint num)
 {
 	if (0 >= num)
@@ -16,8 +16,7 @@ double* OmpNumRandomGenerate(double min, double max, uint num)
 	}
 	srand((unsigned int)time(NULL));
 	double* result = OmpGetMemoryPool()->OmpAlloc(num);
-//#pragma omp parallel for schedule(static)
-	for (int i = 0; i < num; ++i)
+	for (int i = 0; i < (int)num; ++i)
 	{
 		result[i] = OmpRand(min, max);
 	}
@@ -25,11 +24,11 @@ double* OmpNumRandomGenerate(double min, double max, uint num)
 }
 
 /**
- * Generate random number from min value to max value
- * @param min The minimum value of number
- * @param max The maximum value of number
- * @return The double number
- */
+* Generate random number from min value to max value
+* @param min The minimum value of number
+* @param max The maximum value of number
+* @return The double number
+*/
 double OmpRand(double min, double max)
 {
 	double f = 0;
